refactor: simplify testimonial spotlight component API
Remove complex state management and props in favor of CSS variables
for better performance and simpler usage. Replace multiple event
handlers with hover-based styling using Tailwind group utilities.
Consolidate demo files and update documentation to reflect new
CSS-variable-based customization approach.
