+++ src/app/(app)/(blocks)/components/showcase/page.tsx
@@ -30,7 +30,7 @@ import MiddleTruncationDemo from "@/registry/examples/middle-truncation-demo"
 import ScrollFadeEffectDemo from "@/registry/examples/scroll-fade-effect-demo-01"
 import ShimmeringTextDemo2 from "@/registry/examples/shimmering-text-demo-02"
 import SlideToUnlockDemo1 from "@/registry/examples/slide-to-unlock-demo-01"
-import TestimonialSpotlightDemo from "@/registry/examples/testimonial-spotlight-demo-01"
+import TestimonialSpotlightDemo from "@/registry/examples/testimonial-spotlight-demo"
 import TestimonialsMarqueeDemo1 from "@/registry/examples/testimonials-marquee-demo-01"
 import TextFlipDemo from "@/registry/examples/text-flip-demo"
 import ThemeSwitcherDemo from "@/registry/examples/theme-switcher-demo"
