+++ src/features/portfolio/components/components-showcase.tsx
@@ -14,7 +14,7 @@ import IconSwapDemo from "@/registry/examples/icon-swap-demo"
 import MiddleTruncationDemo from "@/registry/examples/middle-truncation-demo"
 import ShimmeringTextDemo2 from "@/registry/examples/shimmering-text-demo-02"
 import SlideToUnlockDemo1 from "@/registry/examples/slide-to-unlock-demo-01"
-import TestimonialSpotlightDemo from "@/registry/examples/testimonial-spotlight-demo-01"
+import TestimonialSpotlightDemo from "@/registry/examples/testimonial-spotlight-demo"
 import TextFlipDemo from "@/registry/examples/text-flip-demo"
 import ThemeSwitcherDemo from "@/registry/examples/theme-switcher-demo"
 import ThemeToggleEffectDemo from "@/registry/examples/theme-toggle-effect-demo/theme-toggle-effect-demo"
