+++ src/registry/examples/_registry.ts
@@ -315,23 +315,12 @@ export const examples: Registry["items"] = [
     ],
   },
   {
-    name: "testimonial-spotlight-demo-01",
+    name: "testimonial-spotlight-demo",
     type: "registry:example",
     registryDependencies: [getRegistryItemUrl("testimonial-spotlight")],
     files: [
       {
-        path: "examples/testimonial-spotlight-demo-01.tsx",
-        type: "registry:example",
-      },
-    ],
-  },
-  {
-    name: "testimonial-spotlight-demo-02",
-    type: "registry:example",
-    registryDependencies: [getRegistryItemUrl("testimonial-spotlight")],
-    files: [
-      {
-        path: "examples/testimonial-spotlight-demo-02.tsx",
+        path: "examples/testimonial-spotlight-demo.tsx",
         type: "registry:example",
       },
     ],
